1 - Removed Bloodflare mana star drops. 2 - Increased Bloodflare heart drop timer from 3 to 5 seconds. 3 - Moved Revivify, Daedalus Absorb and Absorber healing to Hurt instead of PreHurt. 4 - Reduced various life steal/healing bonuses, made their caps more strict and increased their cooldowns. 5 - Changed Baleful Harvester to be true melee and increased skull velocity and changed the skull AI. 6 - Nerfed RE weapons but made their ammo sold by the Arms Dealer and not drop from bosses. 7 - Adjusted more OOA armors by moving stats from the armor pieces to the set bonuses. 8 - Buffed Relic of Convergence healing, because it wasn't good. 9 - Adjusted DoG phase 2 AI to transition phases at different HP intervals, teleport rift creates a fireball explosion when DoG teleports to it when he's below 60% HP. 10 - Hive Cyst, Perforator Cyst and ??? no longer spawn while the pillars are active. 11 - Nerfed some Malice Mode AIs again and buffed others. 12 - Added new Perforator Hive spore attack, it's scuffed as fuck atm because it's using a placeholder sprite and projectile lol. 13 - Polter movement changed when lining up a charge. 14 - Nerfed Event Horizon homing range and black hole duration. 15 - Nerfed several other projectile homing ranges. 16 - Reduced Venom Staff buff from 66% to 33%.

+++ Projectiles/Boss/AstralFlame.cs
@@ -46,13 +46,13 @@ namespace CalamityMod.Projectiles.Boss
             if (projectile.velocity.X < 0f)
             {
                 projectile.spriteDirection = -1;
-                projectile.rotation = (float)Math.Atan2(-projectile.velocity.Y, -projectile.velocity.X);
-            }
+				projectile.rotation = (float)Math.Atan2((double)-(double)projectile.velocity.Y, (double)-(double)projectile.velocity.X);
+			}
             else
             {
                 projectile.spriteDirection = 1;
-                projectile.rotation = (float)Math.Atan2(projectile.velocity.Y, projectile.velocity.X);
-            }
+				projectile.rotation = (float)Math.Atan2((double)projectile.velocity.Y, (double)projectile.velocity.X);
+			}
 
             Lighting.AddLight(projectile.Center, 0.3f, 0.5f, 0.1f);
